+++ .env
@@ -1,0 +1,1 @@
+VITE_SCROLL_THRESHOLD=90
