+++ src/libs/utils/test.ts
@@ -1,40 +1,3 @@
-export const shop = [
-  { url: '/home', label: 'All Products' },
-  { url: '/about', label: 'Daily Care' },
-  { url: '/services', label: 'By Treatment' },
-  { url: '/contact', label: 'By Skin Type' },
-  { url: '/contact', label: 'By Skin Concern' },
-];
-
-export const Learn = [
-  { url: '/home', label: 'Courses' },
-  { url: '/about', label: 'Professional Peels' },
-  { url: '/services', label: 'Skin Concerns' },
-  { url: '/contact', label: 'Skin Types' },
-  { url: '/contact', label: 'Skin Intelligence' },
-];
-
-export const About = [
-  { url: '/home', label: 'Our Story' },
-  { url: '/about', label: 'Our Peel Legacy' },
-  { url: '/services', label: 'Careers' },
-];
-
-export const CustomerService = [
-  { url: '/about', label: 'Contact Us' },
-  { url: '/services', label: 'FAQ' },
-  { url: '/contact', label: 'Shipping' },
-  { url: '/services', label: 'Returns' },
-];
-
-export const OtherLinks = [
-  { url: '/home', label: 'PCA SKIN International' },
-  { url: '/about', label: 'Privacy Policy' },
-  { url: '/services', label: 'Terms & Conditions' },
-  { url: '/contact', label: 'Do Not Sell My Personal Information' },
-  { url: '/services', label: 'Product excellence' },
-];
-
 export const products = [
   {
     id: 1,
